# docs/api/assignment.openapi.yaml
openapi: 3.0.1
info:
  description: Coursistant LMS REST API contract (OpenAPI 3).
  title: LMS API
  version: 1.0.0
servers:
- description: Servlet context-path
  url: /api
security:
- bearerAuth: []
...
paths:
...
  /v2/courses/{courseId}/assignments/{assignmentId}/rubric:
    get:
      description: Course members with assignment visibility. Students only see Published
        assignment rubrics; posted=false when no rubric has been uploaded.
      operationId: assignmentGetRubric
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND or RUBRIC_NOT_FOUND
      summary: Get current rubric pointer
      tags:
      - AssignmentRubrics
    post:
      description: "Instructor/TA. Multipart parts: required `file` (PDF), optional\
        \ `confirmReplaceAfterGrading` when replacing after grading has started."
      operationId: assignmentUploadRubric
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      - in: query
        name: confirmReplaceAfterGrading
        required: false
        schema:
          type: boolean
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                confirmReplaceAfterGrading:
                  type: boolean
                  description: Required true when grades already exist against a previous
                    rubric
                file:
                  type: string
                  format: binary
                  description: PDF rubric file
              required:
              - file
        required: true
      responses:
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "UNSUPPORTED_FILE_TYPE (non-PDF), FILE_TOO_LARGE"
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: RUBRIC_REPLACE_CONFIRM_REQUIRED
      summary: Upload a PDF rubric
      tags:
      - AssignmentRubrics
  /v2/courses/{courseId}/assignments/{assignmentId}/rubric/download:
    get:
      description: "Course members with assignment visibility. Returns PDF bytes with\
        \ Content-Disposition: attachment. Students cannot access Draft assignment\
        \ rubrics."
      operationId: assignmentDownloadRubric
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
          description: Rubric PDF bytes
          headers:
            Content-Disposition:
              description: attachment; filename="..."; filename*=UTF-8''...
              schema:
                type: string
                example: attachment; filename="rubric.pdf"; filename*=UTF-8''rubric.pdf
              style: simple
            Content-Type:
              description: Typically application/pdf
              schema:
                type: string
                example: application/pdf
              style: simple
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND or RUBRIC_NOT_FOUND
      summary: Download current rubric PDF
      tags:
      - AssignmentRubrics
  /v2/courses/{courseId}/assignments/{assignmentId}/rubric/preview:
    get:
      description: "Course members with assignment visibility. Returns PDF bytes with\
        \ Content-Disposition: inline. Students cannot access Draft assignment rubrics."
      operationId: assignmentPreviewRubric
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
          description: Rubric PDF bytes for inline preview
          headers:
            Content-Disposition:
              description: inline; filename="..."; filename*=UTF-8''...
              schema:
                type: string
                example: inline; filename="rubric.pdf"; filename*=UTF-8''rubric.pdf
              style: simple
            Content-Type:
              description: Typically application/pdf
              schema:
                type: string
                example: application/pdf
              style: simple
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: ASSIGNMENT_NOT_FOUND or RUBRIC_NOT_FOUND
      summary: Preview current rubric PDF inline
      tags:
      - AssignmentRubrics
  /v2/courses/{courseId}/assignments/{assignmentId}/rubric/restore-previous:
    post:
      description: Instructor/TA. Rolls the current pointer back one version. May
        require confirmReplaceAfterGrading when grading already occurred against the
        current rubric.
      operationId: assignmentRestorePreviousRubric
      parameters:
      - in: path
        name: courseId
        required: true
        schema:
          type: integer
          format: int32
      - in: path
        name: assignmentId
        required: true
        schema:
          type: integer
          format: int32
      - in: query
        name: confirmReplaceAfterGrading
        required: false
        schema:
          type: boolean
      responses:
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: "ASSIGNMENT_NOT_FOUND, RUBRIC_NOT_FOUND, RUBRIC_NO_PREVIOUS_VERSION"
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
          description: RUBRIC_REPLACE_CONFIRM_REQUIRED
      summary: Restore previous rubric version
      tags:
      - AssignmentRubrics
...
components:
  schemas:
    ApiErrorResponse:
      type: object
      description: Error envelope rendered by ApiExceptionHandler (data optional).
      properties:
        code:
          type: string
          description: Stable ErrorType.name()
          example: USER_NOT_FOUND
        data:
          type: object
          nullable: true
        message:
          type: string
          example: User Does Not Exist
        status:
          type: integer
          format: int32
          example: 404
        timestamp:
          type: string
          format: date-time
          example: 2026-07-23T10:00:00Z
      required:
      - code
      - message
      - status
      - timestamp
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: "Access token from POST /v1/auth/login (Authorization: Bearer <token>)."
      name: bearerAuth
      scheme: bearer
      type: http